mail access link + renamed user.js to db.js + graceful mongoose exit
Function written that emails a link that contains email ID, secret key
once clicked, front uses it along with userID to posts emailID, secret and userID to /mail/verifyFormAccess.
Server checks which form it's for and uses it to assign the form to the user.

users don't really need their own controller. 1 controller for DB access should be enough. Feel free to change if needed.